% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_gmail.R
\name{connect_gmail}
\alias{connect_gmail}
\title{Load the SMTP connection variables for sending email via Gmail.}
\usage{
connect_gmail()
}
\description{
This function is called from WaveR::use_wave()
}
\examples{
\dontrun{
connect_gmail()

}#'
}
