% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_snowflake.R
\name{connect_snowflake}
\alias{connect_snowflake}
\title{Title}
\usage{
connect_snowflake(dbname = "SCARF", schema = "RAW", api = "odbc")
}
\arguments{
\item{dbname}{Database name}

\item{schema}{Schema name}

\item{api}{ODBC or ADBC}
}
\value{
connection object
}
\description{
Title
}
