% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_supabase.R
\name{connect_supabase}
\alias{connect_supabase}
\title{Set up a Supbabase connection object to the specified CapeGlobal App}
\usage{
connect_supabase(app = NA)
}
\arguments{
\item{app}{Specify the CapeGlobal Application in Supabase ("SCARF", "CLOAK")}
}
\value{
Silently returns a connection object
}
\description{
Requires several environment variables for Supabase (HOST, USER, & PASSWORD)
}
\examples{
\dontrun{
connect_supabase()

connect_supabase(dbname = "CLOAK")

connect_supabase(dbname = "SCARF")

}
}
