% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_secrets.R
\name{get_secrets}
\alias{get_secrets}
\title{get_secrets()}
\usage{
get_secrets(path = "/", env = "dev", verbose = TRUE)
}
\arguments{
\item{path}{Set the desired subdirectory that contains the desired secret(s). Defaults to '/', which returns all secrets}

\item{env}{Infisical Secrets Environment. Defaults to 'dev'.}

\item{verbose}{need to finish this}
}
\value{
Silently adds secrets to the environment
}
\description{
get_secrets()
}
