% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_wave.R
\name{use_wave}
\alias{use_wave}
\title{Use the default Wave settings}
\usage{
use_wave(quiet = FALSE)
}
\arguments{
\item{quiet}{Toggles the display of each setting. Default is FALSE}
}
\value{
Silently adds configuration settings to the environment
}
\description{
Loads config.yml settings and default options for a Wave project.
}
\details{
Specifically, reads the config.yml file if present, displays the addresses in
config$emails_errors, and sets the working directory.
}
\examples{
\dontrun{
use_wave()

use_wave(quiet = TRUE)
}
}
