% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_energy.R
\name{wave_energy}
\alias{wave_energy}
\title{Calculate the wave energy flux}
\usage{
wave_energy(height = NA, period = NA, wind = NA, fetch = NA,
  depth = NA)
}
\arguments{
\item{height}{Significant wave height, in meters.}

\item{period}{Peak wave period, in seconds.}

\item{wind}{Wind speed at 10m, in m/s.}

\item{fetch}{Fetch length, in meters.}

\item{depth}{Water depth, in meters.}
}
\value{
Wave energy flux, in kW/m.
}
\description{
Calculates the wave energy flux (power per meter of wave crest) given
either (1) the significant wave height and peak period or (2) the wind speed
at 10m, fetch length and (optionally) water depth.
}
\details{
Given the significant height (\emph{H}) and peak period (\emph{T}),
the wave energy flux is calculated as: \deqn{\frac{\rho g^2}{64 \pi} H^2 T},
where \eqn{\rho} is the density of water (998 kg/m^3) and \emph{g} is the
acceleration of gravity (9.81 m/s^2).

If both \code{height} and \code{period} are missing, they are estimated from
on the wind speed at 10m (\eqn{U_{10}}) and the fetch length (\emph{F}) as
described in Resio et al. (2003):
\deqn{{U_f}^2 = 0.001 (1.1 + 0.035 U_{10}) {U_{10}}^2} (friction velocity)
\deqn{\frac{g H}{{U_f}^2} = \min (0.0413 \sqrt{\frac{g F}{{U_f}^2}}, 211.5)}
\deqn{\frac{g T}{U_f} = \min (0.651 (\frac{g F}{{U_f}^2})^{1/3}, 239.8)}
If the depth (\emph{d}) is specified, it imposes a limit on the peak period:
\deqn{T_{max} = 9.78 \sqrt{\frac{d}{g}}} (in seconds)
}
\examples{
 # With height and period arguments
 wave_energy(8, 1)

 # With wind, fetch and depth arguments (must be named)
 wave_energy(wind = 12, fetch = 15000, depth = 10)
}
\references{
Resio, D.T., Bratos, S.M., and Thompson, E.F. (2003). Meteorology
 and Wave Climate, Chapter II-2. Coastal Engineering Manual.
 US Army Corps of Engineers, Washington DC, 72pp.
}

