% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_snowflake.R
\name{connect_snowflake}
\alias{connect_snowflake}
\title{Set up a Snowflake connection object to the specified database & schema}
\usage{
connect_snowflake(dbname = "SALESFORCE", schema = "RAW", warehouse = "WH_XS")
}
\arguments{
\item{dbname}{Database name. Will have DB_PREFIX and DB_SUFFIX added from global environment}

\item{schema}{Schema name}

\item{warehouse}{Warehouse name. Default is "WH_XS"}
}
\value{
Silently returns a connection object
}
\description{
Requires several environment variables
}
\examples{
\dontrun{
connect_snowflake()

connect_snowflake(dbname = "RAMP", schema = "CLEAN")

}
}
