% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_supabase.R
\name{connect_supabase}
\alias{connect_supabase}
\title{Title}
\usage{
connect_supabase(app = NA, api = "odbc")
}
\arguments{
\item{app}{Specify the StudioBand Application in Supabase ("SCARF", "CLOAK")}

\item{api}{Specify the database driver (ODBC, ADBC). Default is 'odbc'}
}
\value{
conn
}
\description{
Title
}
