% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fetch_len.R
\name{fetch_len}
\alias{fetch_len}
\title{Calculate the fetch length around a point}
\usage{
fetch_len(p, bearings, shoreline, dmax, spread = 0)
}
\arguments{
\item{p}{Coordinates of the point from which to calculate fetch, in degrees.
Either a vector of two numbers, a matrix of two columns (longitude,
latitude), or a SpatialPoints* object.}

\item{bearings}{Vector of bearings, in degrees.}

\item{shoreline}{SpatialLines* or SpatialPolygons* object representing the
shoreline.}

\item{dmax}{Maximum value of fetch length, returned if there is no land
within a distance of \code{dmax} from a given bearing.}

\item{spread}{Vector of relative bearings (in degrees) for which
to calculate fetch around each main bearing (see details).}
}
\value{
A named vector representing the fetch length for each direction
 given in \code{bearings}.
}
\description{
Given a point, a shoreline layer and a vector of wind directions (bearings),
\code{fetch_len} calculates the distance from point to shore for each bearing.
}
\details{
The fetch length (or fetch) is the distance of open water over which the wind
can blow in a specific direction. Note that bearings represent the direction
from where the wind originates.

The optional \code{spread} argument defines relative directions that are
added to each main bearing to produce a set of sub-bearings. The fetch lengths
calculated for each sub-bearing are averaged with weights proportional to
\code{cos(spread)}. By default, \code{spread = 0} and fetch length is
calculated for the main bearings only.

This function converts all Spatial* objects to WGS84 geographic
(longitude, latitude) coordinates and calculates geodesic distances using
the Vincenty ellipsoid formula (from the geosphere R package).

If the shoreline layer is given as SpatialPolygons*, the function verifies
that the input point is outside all polygons (i.e. in water). If this is
not the case, it issues a warning and returns a vector of \code{NA}.
}
\seealso{
\code{\link{fetch_len_proj}} to work in projected coordinates.

\code{\link{fetch_len_multi}} for an efficient alternative when
 computing fetch length for multiple points.
}

