% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fetch_len.R
\name{fetch_len_multi}
\alias{fetch_len_multi}
\title{Calculate the fetch length for multiple points}
\usage{
fetch_len_multi(pts, bearings, shoreline, dmax, spread = 0)
}
\arguments{
\item{pts}{Coordinates of the points from which to calculate fetch, in degrees.
Either a matrix of two columns (longitude, latitude) or a SpatialPoints* object.}

\item{bearings}{Vector of bearings, in degrees.}

\item{shoreline}{SpatialLines* or SpatialPolygons* object representing the
shoreline.}

\item{dmax}{Maximum value of fetch length, returned if there is no land
within a distance of \code{dmax} from a given bearing.}

\item{spread}{Vector of relative bearings (in degrees) for which
to calculate fetch around each main bearing (see \code{\link{fetch_len}}
for details).}
}
\value{
A matrix of fetch lengths, with one row by point in \code{pts} and
 one column by bearing in \code{bearings}.
}
\description{
\code{fetch_len_multi} provides an efficient method to compute fetch length
for multiple points.
}
\details{
This function clips the \code{shoreline} layer to a rectangle around each
 point in \code{pts} before applying \code{\link{fetch_len}} to all points.
 This saves computation time if points of interest are spatially clustered
 and their rectangular buffers overlap.
}

